ref(ts): Roughly type remaining tablechart components
